requirements.txt
webob==1.8.7
parse==1.20.2
requests==2.32.3
requests-wsgi-adapter==0.4.1
pytest==8.2.2
pytest-cov==5.0.0
Jinja2==3.1.4
whitenoise==6.7.0
# for pypi upload
twine==5.1.1
wheel
